add working db and connected with movie project front end

+++ src/main/java/com/codeup/fortran_movies_api/web/MovieDto.java
@@ -1,7 +1,7 @@
 package com.codeup.fortran_movies_api.web;
 
 public class MovieDto {
-    private Integer id;
+    private Long id;
     private String title;
     private String rating;
     private String poster;
@@ -13,7 +13,7 @@ public class MovieDto {
     public MovieDto() {
     }
 
-    public MovieDto(Integer id, String title, String rating, String poster, String year, String genre, String director, String plot) {
+    public MovieDto(Long id, String title, String rating, String poster, String year, String genre, String director, String plot) {
         this.id = id;
         this.title = title;
         this.rating = rating;
@@ -24,11 +24,11 @@ public class MovieDto {
         this.plot = plot;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
